requirements.txt

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0

# Azure AI services
azure-ai-inference==1.0.0b4
azure-core==1.30
azure-ai-documentintelligence==1.0.0b1

# HTTP requests
requests==2.31.0

# JSON and data processing
pathlib2==2.3.7

# Optional: For development and debugging
pytest==7.4.3
pytest-asyncio==0.21.1